refactor(gpu_whir): clippy-clean (0 own-crate warnings)
- Crate-level #![allow(clippy::too_many_arguments)] for the WHIR
  scheduling/launcher functions (same precedent as gpu_hash / gpu_ntt /
  gpu_execution_prover / gpu_trace / gpu_gkr); removed the 10 now-redundant
  per-function allows this superseded (kernels/mod.rs,
  fold/schedule/round_phases.rs, fold/debug.rs).
- upstream.rs: gated the 5 re-exports only used by fold::tests
  (make_eq_poly_in_full, multivariate_coeffs_into_hypercube_evals,
  Blake2sU32MerkleTreeWithCap, ColumnMajorMerkleTreeConstructor,
  BaseFieldQuery) behind #[cfg(test)], matching the gpu_trace/gpu_gkr
  convention (DefaultTreeConstructor/MerkleTreeCapVarLength/WhirCommitment
  stay plain — used in production).
- A cluster of query-parity test scaffolding (query_base_trace_holder_for_
  folded_index, its schedule_* helper, GpuScheduledBaseFieldQuery + decode,
  decode_base_leaf_values, fold::bitreverse_index,
  schedule_reduce_outputs_readback, debug::schedule_special_three_point_
  eval_device, lib.rs's copy_coset_values) was compiled unconditionally but
  only ever called from #[cfg(test)] code (fold::tests, fold::tests::
  query_tests, lib.rs's own test mod) — gated the whole chain (plus their
  now-test-only imports) behind #[cfg(test)] instead of leaving it dead in
  the production build.
- fold/mod.rs: dropped 2 genuinely-unused #[cfg(test)] imports
  (PhantomData, Digest — debug.rs already imports both itself) and 2 names
  (BaseFieldQuery, DefaultTreeConstructor) from a "permanently imported for
  debug.rs" re-export block whose rationale had gone stale (debug.rs now
  imports both directly).
- Two type aliases factored for type_complexity (debug.rs); needless_borrow
  (double `&mut` after `split_at_mut`/`from_raw_parts_mut`, same pattern as
  gpu_trace/gpu_gkr); unnecessary_cast (trailing_zeros()/already-typed
  params); let_and_return; redundant_closure; len_zero; vec_init_then_push;
  useless_vec + unnecessary_mut_passed in a test; field_reassign_with_default.

e2e parity pair (run_add_sub_proof_parity_test{,_sec100}) re-run and PASS
since production files were touched throughout (fold/mod.rs,
fold/schedule/{mod,round_phases}.rs, lib.rs, kernels/mod.rs, pow.rs).

+++ gpu/whir/src/kernels/mod.rs
@@ -13,6 +13,8 @@ use gpu_core::primitives::field::{BF, E4};
 use gpu_ops::simple::pow;
 // Permanently compiled (Task 10): used by the promoted `debug.rs` builders.
 use crate::upstream::FieldExtension;
+// Only used by the #[cfg(test)] `pack_rows_for_whir_leaves_multi_coset` below.
+#[cfg(test)]
 use gpu_core::primitives::utils::GetChunksCount;
 use gpu_core::primitives::utils::{
     get_grid_block_dims_for_threads_count, get_grid_block_dims_for_warp_groups, WARP_SIZE,
@@ -177,7 +179,6 @@ pub(crate) fn accumulate_whir_base_columns(
 /// Fused `accumulate_whir_base_columns` + `serialize_whir_e4_columns`: writes
 /// the E4 result into `result` and the column-major BF vectorization (4
 /// columns of `rows` BFs each) into `serialized_bf` in a single pass.
-#[allow(clippy::too_many_arguments)]
 pub(crate) fn accumulate_whir_base_columns_with_serialized_bf(
     memory_values: &(impl DeviceMatrixChunkImpl<BF> + ?Sized),
     witness_values: &(impl DeviceMatrixChunkImpl<BF> + ?Sized),
@@ -457,7 +458,7 @@ pub(crate) fn partially_evaluate_monomials_by_ref(
             monomials,
             partial_evals,
             z_ptr,
-            log_count as i32,
+            log_count,
         );
         PartiallyEvaluateMonomialFormByRefSmallFunction(
             ab_partially_evaluate_monomial_form_by_ref_small_kernel,
@@ -476,7 +477,7 @@ pub(crate) fn partially_evaluate_monomials_by_ref(
         partial_evals,
         z_ptr,
         z_adjustment_ptr,
-        log_count as i32,
+        log_count,
     );
     PartiallyEvaluateMonomialFormByRefFunction(ab_partially_evaluate_monomial_form_by_ref_kernel)
         .launch(&config, &args)?;
@@ -718,7 +719,6 @@ pub(crate) fn batched_eq_factor_scratch_lens(num_queries: usize) -> (usize, usiz
 /// Builds per-query factored-eq slabs and folds
 /// `sum_q( eq(point_q, gid) * challenges[q] )` into `eq_poly[gid]` (RMW).
 /// Two launches regardless of `num_queries`.
-#[allow(clippy::too_many_arguments)]
 pub(crate) fn launch_batched_accumulate_eq_samples(
     claim_points: *const E4,
     challenges: *const E4,
@@ -818,7 +818,6 @@ fn launch_build_split_eq_table(
 /// using one E4 mul + one E4 add per query in the inner loop.
 /// Three launches total (one per slab build + accumulator), regardless of
 /// `num_queries`.
-#[allow(clippy::too_many_arguments)]
 pub(crate) fn launch_split_accumulate_eq_samples(
     claim_points: *const E4,
     challenges: *const E4,
